requirements.txt
openai>=1.14.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.26
boto3>=1.34.0
streamlit>=1.31.0
plotly>=5.18.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
moto>=5.0.0
